backend/requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
